% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adopt_m49_for_FAO.R
\name{adopt_m49_for_FAO}
\alias{adopt_m49_for_FAO}
\title{adopt_m49_for_FAO}
\usage{
adopt_m49_for_FAO(raw_m49)
}
\arguments{
\item{raw_m49}{is the dataframe returned by scrape_m49.R}
}
\value{
Dataframe containing m49 classifications with same column names as 
the SDG data files, and modifications for FAO.
}
\description{
This is function which takes as an input a dataframe containing the raw M49 
classification scraped from UNSD, and outputs a dataframe with renamed columns for 
consistency with SDG as well as modifications used by FAO.
}
