% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{cleanData
This is function which takes as an input a dataframe containing data
for a particular indicator, and outputs a dataframe with simplied column names
and classes.}
\usage{
cleanData(sdgdf)
}
\arguments{
\item{sdgdf}{is a data frame referring to a specific SDG}
}
\value{
This function returns a dataframe containing the cleaned data frame
}
\description{
cleanData
This is function which takes as an input a dataframe containing data
for a particular indicator, and outputs a dataframe with simplied column names
and classes.
}
