% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(sdg = c("2.1.1", "2.1.2", "2.3.1", "2.3.2", "2.5.1_Plants",
  "2.5.1_Animals", "2.5.2", "2.a.1", "2.c.1", "6.4.1", "6.4.2", "14.4.1",
  "14.6.1", "14.b.1", "15.1.1", "15.2.1", "15.4.2", "15.6.1"),
  source = c("web", "local"))
}
\arguments{
\item{sdg}{is a specific SDG indicator under FAO custodianship}

\item{source}{is the source from which you would like to load the data. If it's
"local" you risk it not being up-to-date. The files disseminated over the "web" 
should always be up-to-date.}
}
\value{
This function returns a tibble containing the latest published data
}
\description{
This is function which takes as an input an SDG indicator, and returns
the latest dataset for that indicator as a dataframe. Pay attention to the naming of the
SDG indicators in the default settings.
}
