% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasets.R
\name{getDatasets}
\alias{getDatasets}
\title{getDatasets}
\usage{
getDatasets(sdg_codes, all = FALSE)
}
\arguments{
\item{all}{is a boolean value equal to TRUE only if the user wants to download all
of the SDG datasets.}

\item{sdgs_codes}{is a vector of codes corresponding to the SDG indicators
for which the user would like to download data.}
}
\value{
The functions returns a list wherein each element is a tibble corresponding
to the datasets requested in the input.
}
\description{
This function takes a vector of SDG codes and returns a list inwhich
each element is the raw SDG data downloaded from the web.
}
