% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R, R/harmonizeCols.R
\name{harmonizeCols}
\alias{harmonizeCols}
\title{harmonizeCols}
\usage{
harmonizeCols(sdgdf)

harmonizeCols(sdgdf)
}
\arguments{
\item{sdgdf}{is a data frame referring to a specific SDG}

\item{sdgdf}{is a data frame referring to a specific SDG}
}
\value{
This function returns a dataframe containing the cleaned data frame

This function returns a dataframe containing the cleaned data frame
}
\description{
This is function which takes as an input a dataframe containing data
for a particular indicator, and outputs a dataframe with SDMX column names and
all columns in the class character.

This is function which takes as an input a dataframe containing data
for a particular indicator, and outputs a dataframe with SDMX concepts as column names and
followed by an attribute name. Example: REF_REA
}
