% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeGeoNames.R
\name{harmonizeGeoNames}
\alias{harmonizeGeoNames}
\title{harmonizeGeoNames}
\usage{
harmonizeGeoNames(cleanSdgs)
}
\arguments{
\item{cleanSdgs}{is a data frame referring one or more SDGs}
}
\value{
This function returns a dataframe containing the cleaned data frame
}
\description{
This is function which takes as an input a dataframe containing data
cleanData, and applies the M49
}
