% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeIndicators.R
\name{harmonizeIndicators}
\alias{harmonizeIndicators}
\title{harmonizeIndicators}
\usage{
harmonizeIndicators(sdgData)
}
\arguments{
\item{sdgData}{a dataframe containing SDG data harvested with this package}
}
\value{
This function returns the dataframe with the indicator column amended
}
\description{
This is function which takes as a dataframe as input containing SDG data, and adds the
word the string ("(complementary)") next to the indicator variable for records which are
complementary SDG indicators and adds indexes to subindicators.
}
