% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeUnits.R
\name{harmonizeUnits}
\alias{harmonizeUnits}
\title{harmonizeUnits}
\usage{
harmonizeUnits(unitcolumn)
}
\arguments{
\item{unitcolumn}{a vector of type character containing inconsistent units of measure}
}
\value{
This function returns a character vector containing the consistent units of measure
}
\description{
This is function which takes as an a column of dataframe (as a vector) which contains units of measure
and returns a vector of the same length using consistent units of measure.
}
