% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_m49.R
\name{scrape_m49}
\alias{scrape_m49}
\title{scrape_m49
This scrapes the m49 geoclassification from the UNSD website. 
The URL of the classification is https://unstats.un.org/unsd/methodology/m49/overview/#.}
\usage{
scrape_m49()
}
\value{
This function returns a dataframe containing the m49 classification
exactly as scraped from the UNSD website.
}
\description{
scrape_m49
This scrapes the m49 geoclassification from the UNSD website. 
The URL of the classification is https://unstats.un.org/unsd/methodology/m49/overview/#.
}
